+++ src/components/productsList/ProductList.jsx
@@ -14,13 +14,14 @@ function ProductList({ data, loading, title }) {
       dispatch(loaderClose());
     }
   }, [loading]);
+
   return (
     <>
       <Container fluid className={Styles.list__con}>
         <Row>
           <h4 className="fw-bold p-3">{title ? title : "Trending Products"}</h4>
           <Col className="d-flex flex-wrap gap-1 row-gap-3">
-            {data &&
+            {data?.length > 0 &&
               data?.map((item, index) => (
                 <ProductCard product={item} key={index} />
               ))}
