+++ src/components/ui/index.js
@@ -1,0 +1,3 @@
+import BuyBtn from "./button/BuyBtn";
+
+export { BuyBtn };
