+++ src/services/api/api.js
@@ -1,13 +1,18 @@
-import axios from "../axios/axios";
+import axiosInt from "../axios/axios";
 
 async function fetchData(query = "", signal = "") {
-  const response = await axios.get(query, {
-    signal: signal,
-  });
-  if (response.status === 200) {
-    return response.data;
-  } else {
-    throw new Error("error on data feching");
+  try {
+    const response = await axiosInt.get(query, {
+      signal: signal,
+    });
+    if (response.status === 200) {
+      return response.data;
+    } else {
+      throw new Error("error on data feching");
+    }
+  } catch (err) {
+    console.log(err);
+    return "failed to load data";
   }
 }
